# malipro/web/novigo-openapi.yaml
openapi: 3.1.0
info:
  title: NOVIGO API
  version: 1.0.0
  summary: Super App du Mali — livraison, marketplace, artisans, wallet et paiements Mobile Money.
  description: |
    # NOVIGO API

    API REST versionnée (`/api/v1`) de la super app **NOVIGO** : livraison à la demande,
    repas, pharmacies, supermarchés, artisans, courses express, marketplace, wallet et
    paiements Mobile Money (Orange Money, Wave).

    ## Authentification
    - **JWT Bearer** (`bearerAuth`) : access token courte durée (15 min) sur toutes les routes protégées.
    - **Refresh Token** (`/auth/refresh-token`) : renouvellement de l'access token (30 jours).
    - **OTP SMS** : flux d'inscription/connexion en deux temps (`/auth/register` → `/auth/verify-otp`).
    - **API Keys internes** (`apiKeyAuth`, header `X-Api-Key`) : communication service-à-service.

    ## RBAC
    Chaque opération porte l'extension `x-required-roles`. Rôles : `CUSTOMER`, `DRIVER`,
    `ARTISAN`, `MERCHANT`, `ADMIN`, `SUPER_ADMIN`, `SUPPORT_AGENT`.

    ## Pagination, tri, filtres
    Les collections acceptent `page`, `limit`, `sort` (ex. `-createdAt`), `search` et des
    filtres spécifiques. Les réponses paginées suivent le schéma `PaginatedResponse` (enveloppe
    `data` + `meta`).

    ## Rate limiting
    Réponses porteuses des en-têtes `X-RateLimit-Limit`, `X-RateLimit-Remaining`,
    `X-RateLimit-Reset`. Le dépassement renvoie `429` avec `Retry-After`.

    ## Temps réel (WebSocket)
...

    ## Enveloppes de réponse
...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token JWT. En-tête `Authorization: Bearer <token>`."
    refreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Refresh token JWT présenté uniquement sur `/auth/refresh-token`.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Clé d'API interne pour la communication service-à-service.
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: Numéro de page (1-indexé).
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      required: false
      description: Taille de page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      required: false
      description: Champ de tri. Préfixe `-` pour l'ordre décroissant (ex. `-createdAt`).
      schema:
        type: string
        examples: ["-createdAt"]
...
    StatusFilterParam:
      name: status
      in: query
      required: false
      description: Filtre par statut.
      schema:
        type: string
    DateFromParam:
      name: dateFrom
      in: query
      required: false
      schema:
        type: string
        format: date-time
    DateToParam:
      name: dateTo
      in: query
      required: false
      schema:
        type: string
        format: date-time
...
  responses:
...
    Unauthorized:
      description: Authentification requise ou token invalide.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Token manquant ou expiré.
              status: 401
    Forbidden:
      description: Droits insuffisants (RBAC).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Accès refusé pour ce rôle.
              status: 403
    NotFound:
      description: Ressource introuvable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Ressource introuvable.
              status: 404
    Conflict:
      description: Conflit d'état (doublon, transition invalide).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: CONFLICT
              message: La ressource existe déjà ou l'état est incompatible.
              status: 409
    UnprocessableEntity:
      description: Erreur de validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Certains champs sont invalides.
              status: 422
              fields:
                - field: phone
                  message: Numéro de téléphone malien invalide.
...
  schemas:
...
    SuccessResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          description: Charge utile.
        message:
          type: [string, "null"]
    PaginationMeta:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
          examples: [1]
        limit:
          type: integer
          examples: [20]
        total:
          type: integer
          examples: [137]
        totalPages:
          type: integer
          examples: [7]
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    PaginatedResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message, status]
          properties:
            code:
              type: string
              examples: ["NOT_FOUND"]
            message:
              type: string
            status:
              type: integer
            traceId:
              type: [string, "null"]
    ValidationErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message, status, fields]
          properties:
            code:
              type: string
              const: VALIDATION_ERROR
            message:
              type: string
            status:
              type: integer
              const: 422
            fields:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # ---------- Primitifs ----------
    GeoPoint:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
          examples: [12.6392]
        lng:
          type: number
          format: double
          examples: [-8.0029]
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: integer
          description: Montant en unité mineure (FCFA, sans décimale).
          examples: [2500]
        currency:
          type: string
          const: XOF

    # ---------- Enums ----------
...
    OrderType:
      type: string
      enum: [FOOD, PHARMACY, GROCERY, PARCEL, ARTISAN_SERVICE, MARKETPLACE]
    OrderStatus:
      type: string
      enum: [PENDING, CONFIRMED, PREPARING, READY, ASSIGNED, PICKED_UP, IN_TRANSIT, DELIVERED, CANCELLED, REFUNDED]
...
    PaymentMethod:
      type: string
      enum: [ORANGE_MONEY, WAVE, CARD, WALLET, CASH]
...
    TransactionType:
      type: string
      enum: [DEPOSIT, WITHDRAWAL, TRANSFER, PAYMENT, REFUND, COMMISSION, PAYOUT]
...
    Address:
      type: object
      required: [id, label, line1, city, location]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
          examples: ["Domicile"]
        line1:
          type: string
        line2:
          type: [string, "null"]
        city:
          type: string
          examples: ["Bamako"]
        district:
          type: [string, "null"]
          examples: ["Hamdallaye ACI 2000"]
        location:
          $ref: '#/components/schemas/GeoPoint'
        isDefault:
          type: boolean
...
    OrderItem:
      type: object
      required: [productId, name, quantity, unitPrice]
      properties:
        productId:
          type: string
          format: uuid
        name:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          $ref: '#/components/schemas/Money'
    Order:
      type: object
      required: [id, customerId, type, status, items, total, createdAt]
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
          examples: ["MLP-2026-000137"]
        customerId:
          type: string
          format: uuid
        storeId:
          type: [string, "null"]
          format: uuid
        type:
          $ref: '#/components/schemas/OrderType'
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        deliveryAddress:
          $ref: '#/components/schemas/Address'
        subtotal:
          $ref: '#/components/schemas/Money'
        deliveryFee:
          $ref: '#/components/schemas/Money'
        total:
          $ref: '#/components/schemas/Money'
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        createdAt:
          type: string
          format: date-time
...
    Transaction:
      type: object
      required: [id, walletId, type, amount, createdAt]
      properties:
        id:
          type: string
          format: uuid
        walletId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/TransactionType'
        amount:
          $ref: '#/components/schemas/Money'
        balanceAfter:
          $ref: '#/components/schemas/Money'
        reference:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
...
    Coupon:
      type: object
      required: [code, discountType, value]
      properties:
        code:
          type: string
          examples: ["MALI10"]
        discountType:
          type: string
          enum: [PERCENT, FIXED]
        value:
          type: integer
          examples: [10]
        minOrderAmount:
          type: [integer, "null"]
        expiresAt:
          type: [string, "null"]
          format: date-time
        isActive:
          type: boolean
...
    CreateOrderRequest:
      type: object
      required: [type, items, deliveryAddressId, paymentMethod]
      properties:
        type:
          $ref: '#/components/schemas/OrderType'
        storeId:
          type: [string, "null"]
          format: uuid
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: string
                format: uuid
              quantity:
                type: integer
                minimum: 1
        deliveryAddressId:
          type: string
          format: uuid
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        couponCode:
          type: [string, "null"]
        note:
          type: [string, "null"]
...
    WalletDepositRequest:
      type: object
      required: [amount, method, phone]
      properties:
        amount:
          type: integer
          minimum: 100
        method:
          $ref: '#/components/schemas/PaymentMethod'
        phone:
          type: string
    WalletTransferRequest:
      type: object
      required: [toPhone, amount]
      properties:
        toPhone:
          type: string
        amount:
          type: integer
          minimum: 100
        note:
          type: [string, "null"]
...
paths:
...
  /customers/me/dashboard:
    get:
      tags: [Customers]
      summary: Tableau de bord client
      operationId: customerDashboard
      x-required-roles: [CUSTOMER]
      responses:
        '200': { description: Données du tableau de bord., content: { application/json: { schema: { $ref: '#/components/schemas/SuccessResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /customers/me/orders:
    get:
      tags: [Customers]
      summary: Mes commandes
      operationId: customerOrders
      x-required-roles: [CUSTOMER]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/StatusFilterParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Commandes paginées.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'
...
  /orders:
    get:
      tags: [Orders]
      summary: Historique des commandes
      operationId: listOrders
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/StatusFilterParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/DateFromParam'
        - $ref: '#/components/parameters/DateToParam'
      responses:
        '200':
          description: Commandes paginées.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'
    post:
      tags: [Orders]
      summary: Créer une commande
      operationId: createOrder
      x-required-roles: [CUSTOMER]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              type: FOOD
              storeId: 6f1c2a90-1b2c-4d3e-8f90-abcdef012345
              items:
                - productId: 11111111-1111-1111-1111-111111111111
                  quantity: 2
              deliveryAddressId: 22222222-2222-2222-2222-222222222222
              paymentMethod: ORANGE_MONEY
              couponCode: MALI10
      responses:
        '201':
          description: Commande créée.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Order'
              example:
                success: true
                data:
                  id: 33333333-3333-3333-3333-333333333333
                  reference: MLP-2026-000137
                  type: FOOD
                  status: PENDING
                  total: { amount: 5250, currency: XOF }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
  /wallet/deposit:
    post:
      tags: [Wallet]
      summary: Alimenter le wallet
      operationId: walletDeposit
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WalletDepositRequest' }
      responses:
        '202': { description: Dépôt initié., content: { application/json: { schema: { $ref: '#/components/schemas/SuccessResponse' } } } }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /wallet/withdraw:
    post:
      tags: [Wallet]
      summary: Retirer du wallet
      operationId: walletWithdraw
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WalletDepositRequest' }
      responses:
        '202': { description: Retrait initié., content: { application/json: { schema: { $ref: '#/components/schemas/SuccessResponse' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /wallet/transfer:
    post:
      tags: [Wallet]
      summary: Transférer vers un autre utilisateur
      operationId: walletTransfer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WalletTransferRequest' }
      responses:
        '200':
          description: Transfert effectué.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Transaction'
        '409': { $ref: '#/components/responses/Conflict' }
  /wallet/transactions:
    get:
      tags: [Wallet]
      summary: Historique des transactions
      operationId: walletTransactions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/TransactionType' }
      responses:
        '200':
          description: Transactions paginées.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Transaction'

  # ============================= CHAT =============================
...
  /promotions/coupons:
    get:
      tags: [Promotions]
      summary: Lister les coupons
      operationId: listCoupons
      responses:
        '200':
          description: Coupons.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Coupon'
    post:
      tags: [Promotions]
      summary: Créer un coupon
      operationId: createCoupon
      x-required-roles: [ADMIN, SUPER_ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Coupon' }
      responses:
        '201': { description: Coupon créé., content: { application/json: { schema: { $ref: '#/components/schemas/SuccessResponse' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /promotions/coupons/validate:
    post:
      tags: [Promotions]
      summary: Valider un coupon
      operationId: validateCoupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
                orderAmount: { type: integer }
      responses:
        '200':
          description: Coupon valide.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Coupon'
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...